refactor(server): Switch to TS #44
