+++ server/package.json
@@ -2,9 +2,11 @@
   "name": "octocake-server",
   "version": "0.0.0",
   "private": true,
-  "type": "module",
+  "main": "dist/index.js",
   "scripts": {
-    "dev": "nodemon ./bin/www",
+    "build": "tsc",
+    "start": "tsc & node .",
+    "dev": "tsc -w & nodemon .",
     "prisma:studio": "yarn prisma studio",
     "prisma:migrate:dev": "yarn prisma migrate dev",
     "prisma:migrate:reset": "yarn prisma migrate reset",
@@ -27,7 +29,18 @@
     "slugify": "^1.6.0"
   },
   "devDependencies": {
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cors": "^2.8.12",
+    "@types/debug": "^4.1.7",
+    "@types/express": "^4.17.13",
+    "@types/http-errors": "^1.8.1",
+    "@types/jsonwebtoken": "^8.5.4",
+    "@types/morgan": "^1.9.3",
+    "@types/node": "^16.4.8",
+    "@types/passport": "^1.0.7",
+    "@types/passport-github2": "^1.2.5",
     "nodemon": "^2.0.12",
-    "prisma": "^2.28.0"
+    "prisma": "^2.28.0",
+    "typescript": "^4.3.5"
   }
 }
