--- server/src/config/prisma.js
@@ -1,7 +1,0 @@
-import Prisma from "@prisma/client";
-
-const { PrismaClient } = Prisma;
-
-export const prisma = new PrismaClient({
-  rejectOnNotFound: { findUnique: true },
-});
