+++ server/src/config/prisma.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "@prisma/client";
+
+export const prisma = new PrismaClient({
+  rejectOnNotFound: true,
+});
