+++ server/src/middlewares/verifyToken.ts
@@ -1,10 +1,13 @@
+import { NextFunction, Request, Response } from "express";
+
 import jwt from "jsonwebtoken";
 
-import { jwt_key } from "../config/credentials.js";
+import { jwt_key } from "../config/credentials";
 
-const verifyToken = (req, res, next) => {
+const verifyToken = (req: Request, res: Response, next: NextFunction) => {
   const token = req.headers["oc_token"];
 
+  // @ts-ignore
   jwt.verify(token, jwt_key, (err, data) => {
     if (err) {
       return null;
