--- server/src/routes/oc_token.js
@@ -1,9 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", (req, res) => {
-  res.send(req.cookies.oc_token);
-});
-
-export default router;
