+++ server/src/routes/oc_token.ts
@@ -1,0 +1,9 @@
+import express, { Request, Response } from "express";
+
+const router = express.Router();
+
+router.get("/", (req: Request, res: Response) => {
+  res.send(req.cookies.oc_token);
+});
+
+export default router;
